rename dynamic env
