+++ environment_dqn_nav/assets/star.png
